requirements/common.txt
Django>=5.0.6
djangorestframework>=3.15.2
environs[django]>=11.0.0
tzdata>=2024.1
drf-standardized-errors>=0.14.0
django-rest-registration>=0.8.3
django-cors-headers>=4.4.0
